Modularizes Sticker Loading into StickersManager
